backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.10.0
